LPS-69343 Split DataFactory to several *DataFactory
