+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/counters.ftl
@@ -1,9 +1,9 @@
-<#assign counterModels = dataFactory.newCounterModels() />
+<#assign counterModels = counterDataFactory.newCounterModels() />
 
 <#list counterModels as counterModel>
 	<#if '${counterModel.name}' == 'com.liferay.counter.kernel.model.Counter'>
 		update Counter set currentId = ${counterModel.currentId} where name = '${counterModel.name}';
 	<#else>
-		${dataFactory.toInsertSQL(counterModel)}
+		${initContext.toInsertSQL(counterModel)}
 	</#if>
 </#list>
