+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/default_dl_file_type.ftl
@@ -1,7 +1,7 @@
-${dataFactory.toInsertSQL(dataFactory.defaultDLFileEntryTypeModel)}
+${initContext.toInsertSQL(dLDataFactory.defaultDLFileEntryTypeModel)}
 
 <@insertDDMStructure
-	_ddmStructureLayoutModel=dataFactory.defaultDLDDMStructureLayoutModel
-	_ddmStructureModel=dataFactory.defaultDLDDMStructureModel
-	_ddmStructureVersionModel=dataFactory.defaultDLDDMStructureVersionModel
+	_ddmStructureLayoutModel=dLDataFactory.defaultDLDDMStructureLayoutModel
+	_ddmStructureModel=dLDataFactory.defaultDLDDMStructureModel
+	_ddmStructureVersionModel=dLDataFactory.defaultDLDDMStructureVersionModel
 />
