+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/default_user.ftl
@@ -1,19 +1,19 @@
 <#-- Default user -->
 
-<@insertUser _userModel=dataFactory.defaultUserModel />
+<@insertUser _userModel=userDataFactory.defaultUserModel />
 
 <#-- Guest user -->
 
-<#assign userModel = dataFactory.guestUserModel />
+<#assign userModel = userDataFactory.guestUserModel />
 
 <@insertGroup
-	_groupModel=dataFactory.newGroupModel(userModel)
+	_groupModel=userDataFactory.newGroupModel(userModel)
 	_publicPageCount=0
 />
 
 <#assign
-	groupIds = [dataFactory.guestGroupModel.groupId]
-	roleIds = [dataFactory.administratorRoleModel.roleId]
+	groupIds = [userDataFactory.guestGroupModel.groupId]
+	roleIds = [userDataFactory.administratorRoleModel.roleId]
 />
 
 <@insertUser
@@ -25,13 +25,13 @@
 <#-- Sample user -->
 
 <#assign
-	userModel = dataFactory.sampleUserModel
+	userModel = userDataFactory.sampleUserModel
 
 	sampleUserId = userModel.userId
 
-	userGroupModel = dataFactory.newGroupModel(userModel)
+	userGroupModel = userDataFactory.newGroupModel(userModel)
 
-	layoutModel = dataFactory.newLayoutModel(userGroupModel.groupId, "home", "", "")
+	layoutModel = layoutDataFactory.newLayoutModel(userGroupModel.groupId, "home", "", "")
 />
 
 <@insertLayout _layoutModel=layoutModel />
@@ -42,8 +42,8 @@
 />
 
 <#assign
-	groupIds = dataFactory.getSequence(dataFactory.maxGroupCount)
-	roleIds = [dataFactory.administratorRoleModel.roleId, dataFactory.powerUserRoleModel.roleId, dataFactory.userRoleModel.roleId]
+	groupIds = counterDataFactory.getSequence(initContext.maxGroupsCount)
+	roleIds = [userDataFactory.administratorRoleModel.roleId, userDataFactory.powerUserRoleModel.roleId, userDataFactory.userRoleModel.roleId]
 />
 
 <@insertUser
@@ -53,7 +53,7 @@
 />
 
 <#list groupIds as groupId>
-	${dataFactory.toInsertSQL(dataFactory.newBlogsStatsUserModel(groupId))}
+	${initContext.toInsertSQL(blogDataFactory.newBlogsStatsUserModel(groupId))}
 
-	${dataFactory.toInsertSQL(dataFactory.newMBStatsUserModel(groupId))}
+	${initContext.toInsertSQL(messageBoardDataFactory.newMBStatsUserModel(groupId))}
 </#list>
