+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/groups.ftl
@@ -1,18 +1,18 @@
-<#assign layoutModel = dataFactory.newLayoutModel(dataFactory.guestGroupModel.groupId, "welcome", "com_liferay_login_web_portlet_LoginPortlet,", "com_liferay_hello_world_web_portlet_HelloWorldPortlet,") />
+<#assign layoutModel = layoutDataFactory.newLayoutModel(userDataFactory.guestGroupModel.groupId, "welcome", "com_liferay_login_web_portlet_LoginPortlet,", "com_liferay_hello_world_web_portlet_HelloWorldPortlet,") />
 
 <@insertLayout _layoutModel=layoutModel />
 
 <@insertGroup
-	_groupModel=dataFactory.globalGroupModel
+	_groupModel=userDataFactory.globalGroupModel
 	_publicPageCount=1
 />
 
 <@insertGroup
-	_groupModel=dataFactory.guestGroupModel
+	_groupModel=userDataFactory.guestGroupModel
 	_publicPageCount=1
 />
 
-<#list dataFactory.groupModels as groupModel>
+<#list userDataFactory.groupModels as groupModel>
 	<#assign groupId = groupModel.groupId />
 
 	<#include "asset_publisher.ftl">
@@ -30,24 +30,24 @@
 	<#include "wiki.ftl">
 
 	<@insertDLFolder
-		_ddmStructureId=dataFactory.defaultDLDDMStructureId
+		_ddmStructureId= dLDataFactory.defaultDLDDMStructureId
 		_dlFolderDepth=1
 		_groupId=groupId
 		_parentDLFolderId=0
 	/>
 
-	<#assign publicLayoutModels = dataFactory.newPublicLayoutModels(groupId) />
+	<#assign publicLayoutModels = layoutDataFactory.newPublicLayoutModels(groupId) />
 
 	<#list publicLayoutModels as publicLayoutModel>
 		<@insertLayout _layoutModel=publicLayoutModel />
 	</#list>
 
-	<#assign publicPageCount = publicLayoutModels?size + dataFactory.maxDDLRecordSetCount + dataFactory.maxJournalArticleCount />
+	<#assign publicPageCount = publicLayoutModels?size + initContext.maxDDLRecordSetCount + initContext.maxJournalArticleCount />
 
 	<@insertGroup
 		_groupModel=groupModel
 		_publicPageCount=publicPageCount
 	/>
 
-	${dataFactory.getCSVWriter("repository").write(groupId + ", " + groupModel.name + "\n")}
+	${initContext.getCSVWriter("repository").write(groupId + ", " + groupModel.name + "\n")}
 </#list>
