+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/roles.ftl
@@ -1,3 +1,3 @@
-<#list dataFactory.roleModels as roleModel>
-	${dataFactory.toInsertSQL(roleModel)}
+<#list userDataFactory.roleModels as roleModel>
+	${initContext.toInsertSQL(roleModel)}
 </#list>
