+++ modules/util/portal-tools-sample-sql-builder/src/main/resources/com/liferay/portal/tools/sample/sql/builder/dependencies/users.ftl
@@ -1,15 +1,15 @@
 <#assign
-	groupIds = dataFactory.getNewUserGroupIds(groupModel.groupId)
-	roleIds = [dataFactory.administratorRoleModel.roleId, dataFactory.powerUserRoleModel.roleId, dataFactory.userRoleModel.roleId]
+	groupIds = userDataFactory.getNewUserGroupIds(groupModel.groupId)
+	roleIds = [userDataFactory.administratorRoleModel.roleId, userDataFactory.powerUserRoleModel.roleId, userDataFactory.userRoleModel.roleId]
 
-	userModels = dataFactory.newUserModels()
+	userModels = userDataFactory.newUserModels()
 />
 
 <#list userModels as userModel>
 	<#assign
-		userGroupModel = dataFactory.newGroupModel(userModel)
+		userGroupModel = userDataFactory.newGroupModel(userModel)
 
-		layoutModel = dataFactory.newLayoutModel(userGroupModel.groupId, "home", "", "")
+		layoutModel = layoutDataFactory.newLayoutModel(userGroupModel.groupId, "home", "", "")
 	/>
 
 	<@insertLayout _layoutModel=layoutModel />
